+++ Pokemon/PokemonType.hpp
@@ -1,0 +1,8 @@
+// defining the enum for Pokemon Types
+enum class PokemonType {
+	Grass,
+    Fire,
+    Water,
+    Electric,
+    Normal // added for the default constructor
+};
