+++ Pokemon/Pokemonchoice.hpp
@@ -1,0 +1,8 @@
+// enum classes for the pokemon choice and type
+
+enum class Pokemonchoice {
+    Bulbasaur = 1,
+    Charmander,
+    Squirtle,
+    Pikachu // default choice
+};
